src/tasks_service/requirements.txt
psycopg2-binary>=2.9.9
grpcio==1.40.0
protobuf==3.20
retry>=0.9.2

